move ColumnValues to parser
